requirements.txt
Django>=4.2,<4.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1